requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
dnspython==2.8.0
fastapi==0.126.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.11
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.46.0b1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
murmurhash==1.0.15
narwhals==2.14.0
nltk==3.9.2
numba==0.63.0b1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.2
preshed==3.0.12
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pymongo==4.15.5
PyMySQL==1.1.2
pyparsing==3.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
smart_open==7.5.0
smmap==5.0.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
streamlit==1.52.2
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
wrapt==2.0.1
xgboost==3.1.3
